Refactor: QasTextTruncate
